requirements.txt
PyYAML
psutil
requests
